Move repetetive topic timer job code into base class
