+++ app/jobs/regular/bump_topic.rb
@@ -1,18 +1,8 @@
 # frozen_string_literal: true
 
 module Jobs
-  class BumpTopic < ::Jobs::Base
-
-    def execute(args)
-      topic_timer = TopicTimer.find_by(id: args[:topic_timer_id])
-      return if !topic_timer&.runnable?
-
-      topic = topic_timer.topic
-      if topic.blank?
-        topic_timer.destroy!
-        return
-      end
-
+  class BumpTopic < ::Jobs::TopicTimerBase
+    def execute_timer_action(topic_timer, topic)
       if Guardian.new(topic_timer.user).can_create_post_on_topic?(topic)
         topic.add_small_action(Discourse.system_user, "autobumped", nil, bump: true)
       end
